requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
bitarray==3.6.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
flyerapi==1.2.3
frozenlist==1.8.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
py-socproof==0.1.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyNaCl==1.5.0
python-dotenv==1.1.1
pytoniq-core==0.1.44
requests==2.32.4
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.41
starlette==0.47.2
tonutils==0.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
x25519==0.0.2
yarl==1.22.0
